Initial refactor for hybrid multi-connection model
- Wrap local connection with Codewind lifecycle manager
- Split CLI wrapper into lifecycle and non-lifecycle
- Separate out codewind tree event listener

Need to add Add/Remove connection commands and fix tree context IDs
