+++ dev/package.json
@@ -522,7 +522,7 @@
         "socketio-wildcard": "^2.0.0",
         "ts-node": "^7.0.1",
         "tslint": "^5.18.0",
-        "typescript": "^3.5.3",
+        "typescript": "^3.6.2",
         "vscode": "^1.1.36"
     },
     "dependencies": {
