+++ dev/src/view/TreeItemContext.ts
@@ -1,0 +1,107 @@
+/*******************************************************************************
+ * Copyright (c) 2019 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v20.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+
+// import * as vscode from "vscode";
+
+import Project from "../codewind/project/Project";
+import Connection from "../codewind/connection/Connection";
+
+
+/**
+ * All of these values must match the viewItem regexes in package.nls.json
+ */
+export enum TreeContextValues {
+    // base
+    BASE = "ext.cw",
+    NO_PROJECTS = "noProjects",
+
+    // Codewind status
+    CW_STOPPED = "cwstatus.stopped",
+    // CW_STARTING = "cwstatus.starting",
+    CW_STARTED = "cwstatus.started",
+
+    // Connection
+    CONN_CONNECTED = "connection.connected",
+    CONN_DISCONNECTED = "connection.disconnected",
+
+    // Project
+    PROJ_BASE = "project",
+
+    // (en|dis)abled are mutex
+    PROJ_ENABLED = "enabled",
+    PROJ_DISABLED = "disabled",
+
+    // debuggable, started both imply enabled
+    PROJ_DEBUGGABLE = "debuggable",
+    PROJ_STARTED = "started",
+
+    // auto build statuses are mutex
+    PROJ_AUTOBUILD_ON = "autoBuildOn",
+    PROJ_AUTOBUILD_OFF = "autoBuildOff",
+
+    PROJ_RESTARTABLE = "restartable",
+    PROJ_METRICS = "metricsAvailable",
+}
+
+export function getConnectionContext(connection: Connection): string {
+    let contextValue: TreeContextValues;
+    if (connection.isConnected) {
+        contextValue = TreeContextValues.CONN_CONNECTED;
+    }
+    else {
+        contextValue = TreeContextValues.CONN_DISCONNECTED;
+    }
+    const cv = buildContextValue([contextValue]);
+    // Log.d(`The context value for ${connection} is ${cv}`);
+    return cv;
+}
+
+export function getProjectContext(project: Project): string {
+    const contextValues: TreeContextValues[] = [ TreeContextValues.PROJ_BASE ];
+
+    if (project.state.isEnabled) {
+        contextValues.push(TreeContextValues.PROJ_ENABLED);
+        if (project.state.isStarted) {
+            contextValues.push(TreeContextValues.PROJ_STARTED);
+        }
+        if (project.state.isDebuggable) {
+            contextValues.push(TreeContextValues.PROJ_DEBUGGABLE);
+        }
+    }
+    else {
+        contextValues.push(TreeContextValues.PROJ_DISABLED);
+    }
+
+    if (project.autoBuildEnabled) {
+        contextValues.push(TreeContextValues.PROJ_AUTOBUILD_ON);
+    }
+    else {
+        contextValues.push(TreeContextValues.PROJ_AUTOBUILD_OFF);
+    }
+
+    if (project.capabilities.supportsRestart) {
+        contextValues.push(TreeContextValues.PROJ_RESTARTABLE);
+    }
+
+    if (project.capabilities.metricsAvailable) {
+        contextValues.push(TreeContextValues.PROJ_METRICS);
+    }
+
+    // The final result will look like eg: "ext.cw.project.enabled.autoBuildOn"
+    const cv = buildContextValue(contextValues);
+    // Log.d(`The context value for ${project.name} is ${cv}`);
+    return cv;
+}
+
+// const CONTEXT_SEPARATOR = ".";
+export function buildContextValue(subvalues: string[]): string {
+    return [ TreeContextValues.BASE, ...subvalues].join(".");
+}
